pyproject.toml
[project]
name = "sistema-academico-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.3",
    "django-allauth>=65.13.1",
    "django>=5.2.8",
    "pyjwt>=2.10.1",
    "pyotp>=2.9.0",
    "qrcode[pil]>=8.2",
    "requests>=2.32.5",
]
